edit resolvers, dtos

+++ backend/src/modules/graphql/services/user.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateQLUser } from '../dto/create-userql.input';
 import { UpdateQLUser } from '../dto/update-user.input';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -7,53 +11,51 @@ import { Repository } from 'typeorm';
 
 @Injectable()
 export class UserQLService {
-
-  constructor(@InjectRepository(QLUser) private readonly qluserRepository: Repository<QLUser>){}
+  constructor(
+    @InjectRepository(QLUser)
+    private readonly qluserRepository: Repository<QLUser>,
+  ) {}
   async create(createQLUser: CreateQLUser) {
-    const user =  this.qluserRepository.create(createQLUser)
-        await this.qluserRepository.save(user)
-        return user
+    const userExist = await this.findOneByEmail(createQLUser.email);
+    if (userExist) {
+      throw new ConflictException('Email already exist');
+    }
+    const user = this.qluserRepository.create(createQLUser);
+    await this.qluserRepository.save(user);
+    return user;
   }
 
   async findAll() {
-    const qlusers = await this.qluserRepository.find()
-        return qlusers
+    const qlusers = await this.qluserRepository.find();
+    return qlusers;
   }
 
   async findOne(id: number) {
-    const post = await this.qluserRepository.findOneBy({id})
-    return post
-   
+    const user = await this.qluserRepository.findOneBy({ id });
+    return user;
+  }
+  async findOneByEmail(email: string) {
+    const user = await this.qluserRepository.findOneBy({ email });
+    return user;
   }
 
-  update(id: number, updateGraphqlInput: UpdateQLUser) {
-    try{
-      this.qluserRepository.update(id,updateGraphqlInput)
-  return true
-  }
-  catch(error){
-      console.log({error})
-      return false
-  }
-  }
+  async update(id: number, updateGraphqlInput: UpdateQLUser) {
+    const userExist = await this.findOne(id);
 
-  remove(id: number) {
-    try{
-      this.qluserRepository.delete(id)
-      return true 
-  }
-  catch(error){
-      console.log({error})
-      return false 
-  }
-  }
-
-  removeByEmail(email: string){
-    try{
-      
+    if (!userExist) {
+      throw new NotFoundException('There is no user with this id');
     }
-    catch(error){
-      console.log({error})
+
+    await this.qluserRepository.update(id, updateGraphqlInput);
+    return userExist;
+  }
+  async remove(id: number) {
+    const userExist = await this.findOne(id);
+    console.log({ userExist });
+    if (!userExist) {
+      throw new NotFoundException('There is no user with this id');
     }
+    await this.qluserRepository.delete(id);
+    return userExist;
   }
 }
